Wait for FIFO availability before write
Change to wait for room in the FIFO buffer before proceeding with the next write.
This fixes the issue of getting "Got [any number smaller than 50000] entries, skipping" when plotting (`plot_last_trace()`). The underlying issue was `get_last_trace()` not receiving the full ADC sampled buffer (50k entries). The reason the full buffer wasn't received, is probably that the host PC is not fast enough to clear the USB buffer, resulting in the Bolt dropping packets when the USB FIFO gets full.